GUI launch via 'ros2 run' command
