--- rover_gui/new_gui/pages_functions/dashboard.py
@@ -1,7 +1,0 @@
-from PyQt5.QtWidgets import QWidget
-from PyQt5 import uic
-
-class Dashboard(QWidget):
-    def __init__(self):
-        super(Dashboard,self).__init__()
-        uic.loadUi("./ui/pages/dashboard.ui", self)
