+++ rover_gui/setup.py
@@ -1,15 +1,18 @@
 from setuptools import find_packages, setup
+import os
+from glob import glob
 
 package_name = 'rover_gui'
 
 setup(
     name=package_name,
     version='0.0.0',
-    packages=find_packages(exclude=['test']),
+    packages=[package_name, 'pages', 'static'],
     data_files=[
         ('share/ament_index/resource_index/packages',
             ['resource/' + package_name]),
         ('share/' + package_name, ['package.xml']),
+        (os.path.join('share', package_name, 'ui'), glob('ui/*')),
     ],
     install_requires=['setuptools'],
     zip_safe=True,
@@ -20,8 +23,7 @@ setup(
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            "main_gui = rover_gui.main_gui:main",
-            "gps_talker = rover_gui.gps_talker:main"
+            "main_gui = rover_gui.main:main"
         ],
     },
 )
