+++ rover_gui/static/icons/arrow-left.png
