+++ rover_gui/static/icons/arrow-right.png
