+++ rover_gui/static/icons/navigation.png
