+++ rover_gui/static/icons/output-onlinepngtools.png
