+++ rover_gui/ui/dashboard.ui
@@ -29,6 +29,27 @@
      </property>
     </widget>
    </item>
+   <item row="1" column="0">
+    <widget class="QCheckBox" name="checkBox">
+     <property name="styleSheet">
+      <string notr="true">QCheckBox::indicator {
+	width: 50px;
+	height: 50px;
+}
+
+QCheckBox::indicator:checked {
+	image: url(:/icons/icons/toggle_on.png);
+}
+
+QCheckBox::indicator:unchecked {
+	image: url(:/icons/icons/toggle_off.png);
+}</string>
+     </property>
+     <property name="text">
+      <string/>
+     </property>
+    </widget>
+   </item>
   </layout>
  </widget>
  <resources/>
