+++ rover_gui/ui/main_window.ui
