+++ rover_gui/ui/navigation.ui
