+++ rover_gui/widgets/toggler.py
@@ -1,0 +1,4 @@
+
+
+#class Toggler(QCheckbox):
+    #def __init__()
